+++ packages/synthlet/src/index.ts
@@ -7,10 +7,12 @@ import { registerClipAmpWorklet } from "@synthlet/clip-amp";
 import { registerClockWorklet } from "@synthlet/clock";
 import { registerDattorroReverbWorklet } from "@synthlet/dattorro-reverb";
 import { registerEuclidWorklet } from "@synthlet/euclid";
-import { registerGraniteWorklet } from "@synthlet/granite/src";
+import { registerGraniteWorklet } from "@synthlet/granite";
 import { registerImpulseWorklet } from "@synthlet/impulse";
 import { registerKarplusStrongWorklet } from "@synthlet/karplus-strong";
+import { registerLevelMeterWorklet } from "@synthlet/level-meter";
 import { registerLfoWorklet } from "@synthlet/lfo";
+import { registerLookaheadLimiterWorklet } from "@synthlet/lookahead-limiter";
 import { registerNoiseWorklet } from "@synthlet/noise";
 import { registerParamWorklet } from "@synthlet/param";
 import { registerPolyblepOscillatorWorklet } from "@synthlet/polyblep-oscillator";
@@ -31,7 +33,9 @@ export * from "@synthlet/euclid";
 export * from "@synthlet/granite";
 export * from "@synthlet/impulse";
 export * from "@synthlet/karplus-strong";
+export * from "@synthlet/level-meter";
 export * from "@synthlet/lfo";
+export * from "@synthlet/lookahead-limiter";
 export * from "@synthlet/noise";
 export * from "@synthlet/param";
 export * from "@synthlet/polyblep-oscillator";
@@ -62,7 +66,9 @@ export function registerAllWorklets(
     registerGraniteWorklet(context),
     registerImpulseWorklet(context),
     registerKarplusStrongWorklet(context),
+    registerLevelMeterWorklet(context),
     registerLfoWorklet(context),
+    registerLookaheadLimiterWorklet(context),
     registerNoiseWorklet(context),
     registerNoiseWorklet(context),
     registerParamWorklet(context),